packages/python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "minions-repos"
version = "0.1.0"
description = "GitHub repo definitions, workflow templates, label sets, and branch protection rules"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["repos","ai","minions"]
authors = [
    { name = "Mehdi Nabhani", email = "mehdi@the-mehdi.com" }
]
dependencies = [
    "minions-sdk>=0.2.1",
]

[project.optional-dependencies]
test = ["pytest>=7.0"]

[project.urls]
Homepage = "https://github.com/mxn2020/minions-repos"
Repository = "https://github.com/mxn2020/minions-repos"

[tool.hatch.build.targets.wheel]
packages = ["minions_repos"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
